requirements.txt
aiogram==2.17.1
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
Deprecated==1.2.13
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.3.0
PyNaCl==1.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
wrapt==1.13.3
yarl==1.7.2
beautifulsoup4~=4.10.0
numpy~=1.22.2
PyDriller~=2.0
PyGithub~=1.55
pytz~=2019.3
scikit-learn~=1.0.2
sklearn==0.0
